Changes by x1cygnus:
just added a harcoded macro Invoke, similar to callpkg except that the
instruction 'Invoke' resolves to a call by default (ti83) and to a RST if
the parameter -plus is specified on the command line.

Changes by dom:
Added in a rudimentary default directory set up (Defined at compile time)
a bit kludgy and not very nice!