Removed Bison parser files (which where a very incomplete work in progress).
Bison grammar not practical due to dual nature of z80asm keywords, i.e.
DEFC NOP = 3 ; define a constant called NOP
NOP ; assemble a NOP instruction
LD A,NOP ; loads A with 3

A Bison grammar requires the NOP to be a terminal token, which complicates
the lexer to be able to return a NAME or a NOP depending on context.