Default stack location for ram model is placed just under the UDGs.
The former default of not changing the stack pointer caused a problem with appmake.  The tap loader generated by appmake CLEARs below the loaded binary which moves SP below the binary.  This caused immediate exit by the program when the heap is automatically generated as the program calculated a negative heap size (bottom of stack - end of bss)