version 2.8.4
BUG_0038: library modules not loaded in sequence
The library modules are now loaded in the sequence given on the command line.

"z80asm -ilib1 -ilib2 obj1.o obj2.o"

now loads obj1, obj2, objects from lib1, objects from lib2 in this order.

The load order was obj1, then all library modules required by obj1, then obj2 and so on.
This posed problems when trying to keep all data at the end of the binary code by defining
a library with all the data and linking it last in the command line - the -i sequence was not
respected and the data appeared in the middle of other library modules.

This also posed problems when trying to override a library function by a user supplied one.