Change the stdio build so that it creates all variants in one go so we don't have to keep rebuilding it over and over again.
Separate the stdio into multiple lists and use them in the target lists.

Add a MODULE directives for the asm files - they are processed using
zcc so they lose the original filename.

Ensure that target specific overrides are first in the lib file then
we don't have to have so much trouble with special cases.

Turn on -DSTANDARDESCAPECHARS for MSX. And doing this ensures that it
works for CP/M as well - the vtstone app which uses (f)gets was failing
due to the wrong character being detected.