For the z88 allow the CRT_MODEL startup option to work as per the new library. This means the data section will be reset on each run from either a raw copy or a compressed copy.
Should be easy to extend to the other ROM targets (ZX, Ace, SMS, MSX, Rex)

Note, in contrast to the the new build, the DATA section is after BSS. This
is needed since the z88 has a split BSS section and some hard requirements
on the placement of the fardata section. Rather than have a spidery mess
of sections, force all builds to be like this.