STLplus was originally intended as a library to extend the STL by providing missing container data structures such as smart-pointers, matrices, trees and graphs.
On the way it became a collection point for other general-purpose
code. The most significant of these is the data persistance functions
which allow any C++ data structure to be made serialisable.

Since v3.0 the library has been split into a collection of 5
libraries. The idea is that some of these can be used stand-alone. For
example, the extension data structures can be used without the rest of
STLplus by just using the containers library.

Download
The source code for the STLplus library collection can be downloaded
from the Sourceforge website:

http://sourceforge.net/projects/stlplus/files/

Download the latest 'stlplus3' project.