Message of previous commit:
Solve bug in z80asm where output files (e.g. sym file) were being deleted in the link step.

When the option -d is passed to z80asm, it only assmbles .asm to .obj if the .obj file does not exist or is older than the asm file.

The sym file was being deleted before checking if the asm file had to be assembled, removing it the link step.