Solve problem with missing symbols when linking a consolidated object file
When consolidating object files, local labels have to be renamed to avoid clashes of labels with the same name in different modules. Both the symbols and all expressions referring to those symbols are affected.

There was an error in the code renaming symbols in expressions that caused it to fail to do the rename in the symbol contained a '_'