Fix #94: Allow functions to be explicitly marked as __stdc so that they are called l2r.
Allow function pointers to be decorated with __smallc, __stdc, __z88dk_callee. These
have real effect and change the code generated __z88dk_fastcall isn't used on function
pointers at this point.

To make this useful, we need to be able to copy the annotation at the point of
assignment so the type of the function pointer mutates depending on what is being
called.