Simplify code generation to be easier to maintain - WORK IN PROGRESS
Generate output files as input is parsed.
Do not generate the binary BMK file; it can be recreated during tests by
reading the asm file comments.
Allow only one ARCH field, as the zilog, rabbit and not_z80 take care of
all the groupings we need.
Remove repetitions in the input file by including the knowledge in the
code generator of which opcode combinations are illegal.