Temporarily revert fix #259
Solve compile failures by reverting to the win32/config.h file.
Change the #ifdef in hist.c from WIN32 to _MSC_VER because gcc on cygwin defines WIN32.
Define the default version string with just the date; to write the compiler version would need a
sequence of #ifdefs for each compiler, as the compiler version is not standardized.