Go back to perl for opcode generation
Restart script to generate parser and test files for the assembler.

Motive to change to back to Perl - contrary to my expectation, Python code is not more readable, at least to my eyes. The "batteries included"
advantage of Python may be also applied to Perl if only standard modules are used, which is done in the new tools/cpu.pl

Motive to restart: in Rabbit the IX-IY swap is not as simple as in Z80 - a couple of instructions use (IX) without prefix, (HL) with 0xDD and (IY) with 0xFD.