Create a z80asm.lib file with routines that might be called by the assembler
The routines used by the assembler while emulating opcodes on a Rabbit (e.g. rrd) are today part of
libsrc\z80_crt0s\z80_emu, and this library has to be included explicitly via -i.

Create a new z80asm.lib file with these routines (for now only rrd and rld) and install it in
${PREFIX}/lib; automatically load it at the end of the options as if the user typed -iz80asm.lib.

Search for the library in the current directory, in the directory of the executable, in executable_dir/../lib
and in ${ZCCCFG}/.. .

A not-found-library is ignored, the user will know from the undefined symbols (e.g. __z80asm__rrd)

Show the included library if the -v option is given.

Note: this was the reason for #326, so that the makefile can call
z80asm -xz80asm.lib -d @dev/z80asm_lib/z80asm_lib.lst
and z80asm_lib.lst only contains the asm file names without a path.