Implemente a DAA emulation routine for Rabbit
DAA can now be used in any target, and resolves to a call __z80asm_daa on a Rabbit.

The implementation is hand-assembled based on the Fuse implementation.

The result of the H flag is not corretly emulated (it fails two of the test cases).
Ticks, which was used to test the implementation also fails the correct emulation of DDA result H flag.

The tests that fail are commented in test.pl.