Fix #409: library make failing on mingw
Problems found and fixed:

1) MinGW transforms the input command line arguments, inserting the full path name of the mingw home
before the given command line argument, i.e. the argument /home/paulo/git/z88dk-msys/libsrc/_DEVELOPMENT/
to lstmanip is received in argv[] as c:/MinGW/msys/1.0/home/paulo/git/z88dk-msys/libsrc/_DEVELOPMENT/

The result is that the $(subts...) in the makefiles that replace the library root NEWLIB_DIRECTORY by obj/z80/
fail to find the library root and the result paths are wrong.

Solved by a very hughly hack in lstmanip - filtering the output and removing what seems to be a MinGW root directory
prefix.

2) In the MinGW build sometimes a double slash is inserted between directories, causing the same $(subst...) as
above to fail to find the library root and the result paths are wrong.

Fixed by filtering the output of lstmanip, removing duplicated slashes, and the same in Make.config.

3) Some makefines did not have the trailing slash in obj/z80, causing in MinGW paths like obj/z80ctype, which then
failed to assemble. Fixed.