Fix #442: support globbing in file names and in list files
Add globbing support to z80asm in the command line args (for the Windows users) and
in list files. The standard '*' and '?' of glob.h are supported.

Added also the extended pattern '**' that expands to all subdirectories from the given point,
i.e. lib/**/*.asm will expand to all asm files in any subdierectory of lib.