Fix #436: add ALIGN directive
Add the ALIGN directive to insert padding until the assembly address is a multiple of the align parameter.
The padding is by default 0x00, but can be changed with --filler=0xNN.

If the ALIGN directive appears before any section code, it will cause the previous section to be padded during linking.

If the ALIGN directive appears anywhere in the section code, it includes padding bytes to align the next ASMPC. Note that in this case, the initial section address is not yet known and is considered 0.