Issue #519: Switch over to using ucpp in this branch.
Remove the old zcpp, fix the vcxproj files (maybe)

Fix up a rule that was overly catching. ucpp preserves whitespace
so was catching some hand written assembler in one of the debug/
routines.

Fix osca/flos/mkdir so that we don't have a macro/func clash.

Turn off mmap in ucpp - it doesn't work on windows and it's marked
as experimental in anycase.