Add msleep() to stdlib that will sleep for so many milliseconds.
Define __CPU_CLOCK in the crt0 for each of the targets so that we don't
have to build it for each platform, also allows it to be overridden at
compile time.

The +pps, +test, +rcm times should be 6MHz which is as far as we can go
using this timing routine.

The rcm one will be way off in anycase since instructions have a different
time resolution.

Compile push_di, pop_ei with the correct cpu flags so need to have the
rabbit variants in there (even though they won't be called)