Fix #655: DEFC constants with expressions based on PHASED labels were being patched
This one was hairy!

IF a DEFC constant (ENTRY in cpm22bios.asm) was defined equal to a label (_cpm_bdos_head)
AND the label (_cpm_bdos_head) was in a PHASED block, therefore a constant and not an address,
THEN the expression value (_cpm_bdos_head) was being patched to the object file at the ASMPC
where it was referred to (DEFC ENTRY = _cpm_bdos_head was at ASMPC = 1, i.e. after the $55)

The $FF was the low byte, the high byte was later overwritten in the object file by the JP COMMAND
argument.

Fixed.

Thanks for reporting!