Create a Dockerfile for z88dk (#676)
* Create a Dockerfile for z88dk

This file creates a container that includes a compiled z88dk git snapshot including a patched SDCC version to be used with Docker.
Note: in SDCC compilation the ds390 and ds400 ports had to be disabled to avoid compile errors. On the other hand I am not sure which other ports can be disabled to have a minimal SDCC compilation to be used in z88dk.

* Update z88dk.Dockerfile

* Disable more ports on SDCC compilation

More SDCC ports can be disabled for z88dk support. Because of this, the gputils package is not needed anymore. Also the Python package was not needed.

* Change z88dk install path inside Docker container

Install z88dk sources inside the container in a more suitable path.
Also provide the possibility of having unpatched SDCC compiled and installed in the Docker container.

* Tweak a command to support Windows execution

Add {} to $PWD so current directory is correctly resolved under Windows Powershell (and other *nix shells).

* Changes to have a tinyer Docker image

* Add the '--squash' parameter to Docker build command.
* Use Alpine Linux as base distro.
* Do a GIT shallow clone of the z88dk repository.
* After copying the neccesary binaries, get rid of SDCC source tree:  keeping it is out of scope of this Dockerfile, and a new one could be created to have a SDCC-only Docker image.
* After z88dk & sdcc compilation, remove unneeded distro packages.

* Make Dockerfile more compact and readable

Just some cosmetic updates

* More Dockerfile polishing

* Remove the '--squash' parameter from the Docker build as it is still experimental.
* Add a Docker_Version to indicate the Docker version in which this Dockerfile was developed and tested.
* Mark more packages to be removed after z88dk compilation.

* Ability to customize the z88dk build

Add a simple mechanism to customize the z88dk build