Fix test scripts to run under MSYS2 on Windows
Under MSYS2 a '@file' is interpreted in command line and the file contents is expanded into argv[], but with a rather small limit of input files.
Under MSYS2 a '${ENV_VAR}' is expanded in the command line, unless the variable does not exist.

Enclose all '@' and '$' arguments in double-quotes before calling z80asm in the test scripts.