+pc6001 make the modes modular, we can exclude the code for a mode with:
-pragma-define:CLIB_DISABLE_MODE2=1 or
-pragma-define:CLIB_DISABLE_MODE1=1

Thus we have the benefit of supporting multiple different screen modes
but can still get a compact binary. The default text mode is always
enabled since the code for that is pretty small.

This can be extended to the graphics mode (where we plot in 3 different
modes so will be a little painful), and should be extended to the
other machines with modes.