Fix #885: bug in calculating expressions with one symbol and other operands
When an expression was defined as a label and some computuation (e.g. label %8)
the computation was done at assembly time, when the label is not yet defined,
causing wrong results.

Moved the computation of these expressions to the link phase.
Moved also similar expressions used as JR arguments to the link phase.
Bumpped the object file version to accomodate the JR argument expression.

One side effect of this fix is that JR to different modules is no longer forbidden,
and the int range error if the jump is too far comes in the link phase.