Issue #772 (ish): Einstein disc writer
Compile +cpm -subtype=einstein and a .dsk will be produced.

This disc can be inserted as secondary disc to the DOS boot disc

In theory takes an option to write a boot image to the first two
tracks of the disc. But I've not managed to create a good image
yet.