Fix #917: Make sure the parameter is the one we actually want.
Check for whitespace after the parameter name.

Change 0dd53b73db627d11dee4ba148ddcd06ca335cd21 introduced the new
map file format. In turn this changed the order of the symbols within
the mapfile.

The z88 application DOR has a lot of symbols with similar names, so
when search for "in_dor" we ended up matching the first one, which
actually didn't correspond to the actual DOR. Thus we ended up
corruption of the DOR chain.