Fix msys2 mingw-64 builds
Write output files in binary mode, so that format is consistent with .gitattributes text=eol, and files are not shown as modified in Windows.

In mingw-64, the Windows path C:\... is translated to /c/... (lower case) - need to convert drive letter to lower case.
(this might break builds with the old msys mingw-32)