Fix Makefiles to pass mingw-64 builds
Uniformize usage of EXESUFFIX instead of EXE in all makefiles; add setting of EXESUFFIX to a few Makefiles where it was missing and caused the build in mingw-32 to fail.

Add a crude rule to the top-level Makefile to build each of the bin/ tools if any of the *.c or *.h in the respective sources are out-of-date. This fixes the problem of working with out-of-date tools fetched from bin/.

Uniformize z80asm Makefile to build on the install: target, so that the top level Makefile only calls the z80asm one once.