Allow 32-bit addresses in z80asm (see #1287)
Extend z80asm to handle 24- or 32-bit addresses, where the lower 16-bits are the address seen by the CPU, and the upper 8- or 16-bits are the bank id (platform dependent, e.g. value to be written to the bank register). CALL, JP, ... need to accept the 32-bit address and ignore the upper 16-bits. For the Spectrum 128K (which I know better than the GameBoy), one could write

  section xxx1  ; name is not relevant
  org $00C000       ; select page 0 at address $C000

  public func1
  func1: ...

  section xxx2
  org $01C000        ; select page 1 at address $C000

  extern func1
  ...
  call banked_call   ; platform-dependent function that switches banks and calls the function
  defp func1         ; patch in a 24-bit address (Spectrum 128k); use defq for a 32-bit address

As a side-effect of this change, z80asm no longer emits warnings for 16-bit values out of range.