Issue #1218: Handle running off the bottom of the screen better.
If we fill the last line (so we nudge onto the condition where
we would scroll when we next print) and then print a lf, we end
up on line console_h+1 i.e. two rows below the max printable line.

So change the condition that checks for below the bottom of the
screen.