Ticks: option without arguments is ignored if it is the last argument
Ticks assumes all options to have arguments, and then fixes argc and argv when an option without argument is parsed.
As a side effect, any option without arguments is ignored if passed as the last command line argument (e.g. in "ticks file.bin -trace", -trace is ignored)

This commit does not fix the error, but changes the usage string to request the file.bin as the last argument.