+dynamic allocation of D-FILE
In pseudo-HRG mode the D-FILE size depends on maxy.
The libray can be rebuilt for different values of this parameter; renaming the resulting files permits an easy insertion of such customizations.
The automatic allocation of the display memory happens ONLY if the value for base_graphics is left at zero, so that static buffers and splash screens can be provided.   To activate a custom splash screen without wiping it, use hrg_on() rather than clg().