Update to newer version of gcc
Configuration was forcing gcc 4.9. This was necessary in the past to allow C11 to be compiled.

Current Travis-CI for C++ uses gcc 5.4, so the setting for gcc 4.9 is no longer necessary. Moreover g++-4.9 was causing a core dump in z80asm2 in the RE-flex library.

Also included, and commented-out: code to show a backtrace after a core dump by calling gdb.