Add possibility do define a numeric value to a -D-defined variable in the command line
Option -Dvariable=value defines the variable with the given numeric value. The value can be passed in decimal (e.g. -Dvar=255) or hexadecimal
(e.g. -Dvar=0xff or -Dvar=0ffh or -Dvar=$ff). Note that the $ may have to be escaped from the shell.

If the value is not given, i.e. -Dvariable, the previous behaviour is kept, i.e. the variable is defined with the value 1.