(msx) MegaROM tweaks.
1. Remove the default mapper. Force the user to decide since I know
   nothing about MSX conventions. This also has the advantage that
   if someone writes a really terrible ROM mapper heuristic that
   doesn't respect ROM size then we won't interfere with it when
   a standard ROM is created.

2. Attempt to bias the mapper detection heuristic to use the
   mapper that the user has selected at compilation time.

The ROM detection code that I've seen so far (Mame, Takeda,
BlueMSX, OpenMSX, CocoaMSX) use:

a. A checksum and a check in a local database to find the
   mapper type (not on Takeda)
b. ROM size detection trigger
c. Simple counting of the number of times ld (XXXX),a occurs,
   BlueMSX,OpenMSX,CocoaMSX skews ASCII8 down slightly.

It looks like fMSX doesn't support mappers (but that emulator is
old and I believe not in regular use).

WebMSX seems to use a checksum model, and then falls back onto
a priority based filesize matcher which runs "ASCII8, ASCII16,
Konami". Thankfully it does have a "Set Rom Type" option to
set the correct mapper.

Given that hand written code may well liberally litter code with
this opcode and we've centralised all banking into a single place
we need to overcome any errant matching sequences in the data section:
We know that compiled code can't generate the magic sequences,
furthermore the z88dk library code doesn't contain it either.

For this biasing to work we use all the addresses supported by the
Mapper even though we don't use them directly. This means that
for the mappers we have the following refcount in the heuristics
(obviously which one depends on user config)

- Konami:  8
- ASCII8:  10 (-1 bias from BlueMSX/OpenMSX)
- ASCII16: 9

Which hopefully will be enough to bias things in the way we want.

3. Create the 256 pages supported by ASCII16.