Fix Perl module build in Azure
PERL5LIB environment variable is used by Perl, so rename the cache dir variable to PERL_CACHE. Had to rename the cache to perl_cache after this change.
Add run on pull-request.
Remove all Perl modules except local::lib from the apt-get install, as the versions in the Ubuntu packages are normally a bit behind. Install them from CPAN instead.
eval... line needs to have -Mlocal::lib=$PERL_CACHE when the local:lib is not in the default ~/perl5, which is the case in Azure. It should also be run before cpanm, to have the environment all set. This line has to be repeated before all .build lines to setup the Perl environment.