(pc88) Checkpoint a chunk of work.
1. Rewrite the graphics code so it's in the library. Default is
hires graphics. To use lores graphics switch mode with
console_ioctl() and add 32

2. Use the ALU for graphic mode gencon

3. Use the ALU for plot/unplot in hires mode

4. Rewrite the attribute handling. This does make text mode really
slow (relatively), but it's now usable without having to make
compensation for the "interesting way" it's implemented.

Everytime we print a char in text mode, we grab the appropriate
attribute row, decompress it, set the attributes for the print
position and then recompress. This allows us to mix lores graphics
and text fairly easily.

5. Lores graphics doesn't work in 40 column mode. I've no idea
what is going on if I'm honest.

TODO:

* vpeek() for hires console
* Fix up putsprite/stencil draw - these should use the ALU to
speed things up nicely
* C header file for pc88 hardware
* asm header file for pc88 hardware
* Fix up everything I've broken