Issue #1709: Support for 64k programs on the sam
- Page VRAM in and out as required (bulk of these changes)
- Hook interrupts correctly - checking status flag
- Add line interrupt vectors
- Don't bother with having samdos as bootstrap, just use option

Allram model has the lower page non changing, so we have to
ensure screen routines + crt0 end up in there, so jiggle
the section map a little to ensure this.

The need for bootstrap loading is nicely avoided by the Sam
loading a file into subsequent banks automagically.

This layout in theory would allow us to have banked code as well.

TODO: Need to have a stack area in lower 32k that's actually
allocated