(sccz80) Remove redundant loads
This is achieved by marking the jp as EOS which means that the
registers are considered unassigned. We can then remove loads
from just before then.

Add in a bunch of optimisation rules for bitfields and other
operations that I spotted sifting through output.

For some projects these changes are quite significant in terms of
size - of course there's a performance improvement as well