(appmake) Fix an issue on windows when -o was specified with classic
If -o was specified and it was a RAM compile, then z80asm created
the output file as specified.

If you then ran through appmake we ended up reading and writing
to the same file and...windows.

Only affected classic since newlib creates file_CODE.bin at
the link stage.