Add inp/outp for the 8080
Uses the same technique as AZTEC C which is to dynamically write the subroutine
into reserved data memory and call/jump to it.  This should allow the code to
remain ROMable.

While here, change outp_callee to use hl instead of af for stack juggling.
af was sending the program off into lala land.