(nabu) Checkpoint the hcca work
Intention is that we use interrupts to both read and write
to the port.

We can either read into a ringbuffer (for shortmessages) or into
a spot in memory.

Add support for headered messages (Type, length, payload) that are
read directly into memory.

Also support for retronet.

All completely WiP unfortunately.