(cpc) Get wyz working properly
When we call the firmware, we have to restore the ROM interrupt
handler.

The interrupt handler is called 6 times per frame, so add in a counter
to deal with that. Howver, when we call the firmware, we disable
our interrupt handler and so our interrupt code runs too slowly.

So, rather than continually poll the keyboard, only poll the keyboard
every vsync or so.