Fix #2260: output dir prepended twice
When a library is being created (-x), and an output directory is being
created (-O), and a file without extension is passed, and the option -d
is used, and the object file is up-to-date in respect to the asm file,
the full object file name with the output dir prepended is returned from
the argument parsing function; this object file is then fed into the
library creating function, who prepends the output dir again.

If an asm file was passed as an argument, then the library creating
funciton did the right thing, prepending the output dir.