Fix #2332 Build fails on cygwin
On cygwin the state machine generator ragel dies with a segmentation fault.
This is related with the size of the state machine, as it started
happening when the ld r, rotatate/bit (ixy+d) instructions were added.

Only generate the state machines with make dev, make all uses the
committed generated files.