Issue #2320 add string tables to object and library files
Object files now have a string table at the end, and all references
to strings are replaced by the index into the table; this reduces
duplication in the object files and allows all data items to be
aligned on a 4-byte address, and the usade of a simple *p to read
the value.

Library files include a string table at the end of all the Public
symbols defined in all its modules; a check on this table is made
at the start of linking and the whole library is skipped if it has
none of the needed extern symbols.