Issue #2320 (z80asm) cleanup parsing of command line arguments
When user supplies a *.o, read object file, issue error if not found or not compatible CPU.
(was searching for a .asm file if .o was not found)

When user supplies a *.asm file, always assemble.
(was using an existing .o file if up-to-date)

When user supplies a basename, try first basename.asm and then basename.o.

With -d, the user may supply either the .asm or .o, and z80asm will use the .o if up-to-date. If the
.o is for a different CPU, an error message is output.

With -m*, -d is not allowed.