Fix #2566: error in expansion of nested macros with same parameter names
The preprocessor was creating a new layer of macro expansion and in that
layer was defining the arguments as new macros. This worked well,
except when macro A has an argument x, and calls macro B that also has
an argument x.

Now a textual replacement of all arguments by the parameters is made
before creating a new layer of macro expansion.