Fix #2587 local labels
Implement local labels with the same use case as sjasmplus:
a local label is prefixed with @ and is valid in a block between
two non-local labels. A local label name cannot be repeated in the
same block, but can be repeated in another block.

The local label full name is <last-non-local-label>@<local-label>.

A local label can be referred to as its short name @label or long name
prev_non_local@label.