Bugfixes
Simplify file open error.
REPT inside a MACRO was not respecting a previous C_LINE.
Error was outputting both source and expanded line even when equal.
RPTC inside a MACRO was getting the input string converted to integers before run.
Unclosed IF error was not reporting the error line number.
Error message of unclosed structure was not reporting start line number of structure.