Add data model for the Assembler/Linker
Still needs to be integrated with existing code.
Add FileCache and classes to read text and binary files
Add tokenizer TokenFileReader that reads from the FileCache and tokenizes
Add line splitting to TokenFileReader
Add tokenized file caching to TokenFileReader, so that a file is only tokenized once
Add the possibility to inject tokens to a TokenFileReader, e.g. a macro expansion can be injected back to the input stream
Add new Expression class with silent parsing and exception throwing on evaluation errors
Add new Symbol class
Reorganize data model in different modules
Add parent to Opcode and Patch to be able to compute asmpc and opcode_size internally
Add Macro creation, parsing and expanding
Add DependencyFile generator
Add REPT, REPTI and REPTC creation and expansion
Add text injection to TokenFileReader, that is tokenized and returned before the rest of the input
Integrate with the new interface of TokenLine
Add pragma once and ifndef guard detector to TokenLineReader
Reduced copies of vectors of TokenLine in managing the cache
Run the include guard detection on open() so that the caller can prevent reading the file altogether if it has already been read
Extend Macros::expand to recurse for chained macros and process the stringize (#) and token paste (##) operators
Refactor Preprocessor