(p2000) Port classic input routines to P2000t
This ports the classic zx81 input routines to the
P2000t. Note that the P2000t usually handles the
keyboard in the ROM, and that the GetKey experience
is likely subpar to what the ROM provides.

Furthermore the shift key is ignored for in_LookupKey,
in_KeyPressed and in_JoyKeyboard.